+++ src/handle/tactic.ml
@@ -203,6 +203,7 @@ type tactic =
   | T_assumption
   | T_change
   | T_fail
+  | T_focus
   | T_generalize
   | T_have
   | T_induction
@@ -236,6 +237,7 @@ let get_config (ss:Sig_state.t) (pos:Pos.popt) : config =
   add "assumption" T_assumption;
   add "change" T_change;
   add "fail" T_fail;
+  add "focus" T_focus;
   add "generalize" T_generalize;
   add "have" T_have;
   add "induction" T_induction;
@@ -340,6 +342,18 @@ let p_rwpatt_of_string (pos:popt) (t:term): p_rwpatt option =
            Some (Parsing.Parser.Lp.parse_rwpatt_string p string)
   | _ -> fatal pos "not a string literal"
 
+(** [p_int_of_string pos t] returns the int contained in a string literal
+    term [t]. *)
+let p_int_of_string : popt -> term -> int = fun pos t ->
+  match t with
+  | Symb s when String.is_string_literal s.sym_name ->
+      begin
+        let string = String.trim (remove_quotes s.sym_name) in
+        try int_of_string string
+        with _ -> fatal pos "string should contain an int"
+      end
+  | _ -> fatal pos "not a string literal"
+
 let is_right (pos:popt) (t:term): bool =
   match t with
   | Symb s when String.is_string_literal s.sym_name ->
@@ -381,6 +395,8 @@ let p_tactic (ss:Sig_state.t) (g:goal) (env:Env.t) (pos:Pos.popt) (t:term)
             | T_change, [_;t] -> P_tac_apply (p_term t)
             | T_change, _ -> assert false
             | T_fail, _ -> P_tac_fail
+            | T_focus, [t] -> P_tac_focus (p_int_of_string pos t)
+            | T_focus, _ -> assert false
             | T_generalize, [_;t] -> P_tac_generalize(p_ident_of_var pos t)
             | T_generalize, _ -> assert false
             | T_have, [t1;t2] ->
@@ -512,6 +528,18 @@ let handle (ss:Sig_state.t) (sym_pos:popt) (priv:bool)
       let id =  Pos.make pos (P_Abst(vparam,idbody)) in
       let pt = Pos.make pos (P_Appl(id,Pos.make pos P_Wild)) in
       tac_refine pos ps gt gs (new_problem()) (scope pt)
+  | P_tac_focus n -> begin
+      let rec extract n l =
+        match l with
+        | [] -> raise Not_found
+        | x::l' -> if n = 0 then x,l'
+                   else let (y,r) = extract (n-1) l' in (y,x::r) in
+      try
+        let (x,gs) = extract n ps.proof_goals in
+        {ps with proof_goals = x::gs}
+      with
+        Not_found -> ps
+    end
   | P_tac_generalize {elt=id; pos=idpos} ->
       (* From a goal [e1,id:a,e2 ⊢ ?[e1,id,e2] : u], generate a new goal [e1 ⊢
          ?m[e1] : Π id:a, Π e2, u], and refine [?[e]] with [?m[e1] id e2]. *)